+++ memor/response.py
@@ -194,14 +194,7 @@ class Response(Message):
 
         :param file_path: response file path
         """
-        result = {"status": True, "message": DATA_SAVE_SUCCESS_MESSAGE}
-        try:
-            with open(file_path, "w") as file:
-                json.dump(self.to_json(), file)
-        except Exception as e:
-            result["status"] = False
-            result["message"] = str(e)
-        return result
+        return self._save_json(file_path, self.to_json())
 
     @staticmethod
     def _validate_extract_json(json_object: Union[str, Dict[str, Any]]) -> Dict[str, Any]: